server/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
ats-resume-scorer
pypdf
python-docx
python-multipart
google-auth
requests
python-dotenv
click<8.1.8